[diffusion] refactor: move model_stages into stages folder
